Convert docstring to comment.
When a docstring is in the middle of the code (instead of at the very top of
it), it doesn't have the effects of being part of the __doc__ property of
the object (function, class etc) and becomes a useless statement (i.e., a
noop).

Just convert that to a real comment.