Guard case of empty element raising exception.
For some reason, with the current BeautifulSoup module from Debian unstable
(3.2.1-1), trying to access the .string of an element that has none (in this
case, `<guid isPermaLink="false"></guid>`) raises an exception instead of
returning an empty string.

This commit adds a test to guard against that kind of problem.